backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Pipecat with providers
pipecat-ai[silero,deepgram,cartesia,groq,webrtc]>=0.0.70

# Asset processing
pdf2image>=1.16.3
Pillow>=10.0.0
PyYAML>=6.0
python-pptx>=0.6.23
PyMuPDF>=1.23.0

# Utilities
aiofiles>=23.2.1
python-dotenv>=1.0.0
